guidance scale in float data type

+++ api.py
@@ -19,7 +19,7 @@ class Request(BaseModel):
     width: int = Field(512, ge=16, le=1024, description="Width of the generated image in pixels. Must can be devided by 16.")
     height: int = Field(512, ge=16, le=1024, description="Height of the generated image in pixels. Must can be devided by 16.")
     num_inference_steps: int = Field(50, ge=1, le=100, description="Number of steps for the image generation process.")
-    guidance_scale: int = Field(3.0, ge=1.0, le=5.0, description="High guidance scales improve prompt adherence.")
+    guidance_scale: float = Field(3.0, ge=1.0, le=5.0, description="High guidance scales improve prompt adherence.")
 
 
 api = FastAPI(title="flux-api", version="0.1.0", docs_url=None, redoc_url=None)
